+++ src/TasksConstraint/RestartQuestionConstraint.php
@@ -3,6 +3,7 @@
 namespace Nuntius\TasksConstraint;
 
 use Nuntius\AbstractQuestionConstraint;
+use Nuntius\Nuntius;
 
 /**
  * Validating the restart question tasks.
@@ -11,12 +12,30 @@ use Nuntius\AbstractQuestionConstraint;
 class RestartQuestionConstraint extends AbstractQuestionConstraint {
 
   /**
+   * Validate the input of tasks.
+   *
+   * @param $value
+   *   The input of the user.
+   *
+   * @return bool|string
+   */
+  public function validateGetTaskId($value) {
+    $tasks = Nuntius::getTasksManager()->getRestartableTasks();
+    foreach ($tasks as $task) {
+      if ($task['label'] == $value) {
+        return TRUE;
+      }
+    }
+    return "Hmmm..... it's look";
+  }
+
+  /**
    * Validate the user input the correct text.
    *
    * @param string $value
    *   The input of the user.
    *
-   * @return bool
+   * @return bool|string
    */
   public function validateStartingAgain($value) {
     if (!in_array($value, ['yes', 'no', 'y', 'n'])) {
